+++ exporters/prometheus/exporter_test.go
@@ -29,13 +29,18 @@ import (
 	"go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/view"
+	"go.opentelemetry.io/otel/sdk/resource"
+	semconv "go.opentelemetry.io/otel/semconv/v1.12.0"
 )
 
 func TestPrometheusExporter(t *testing.T) {
 	testCases := []struct {
-		name          string
-		recordMetrics func(ctx context.Context, meter otelmetric.Meter)
-		expectedFile  string
+		name               string
+		emptyResource      bool
+		customResouceAttrs []attribute.KeyValue
+		recordMetrics      func(ctx context.Context, meter otelmetric.Meter)
+		withoutTargetInfo  bool
+		expectedFile       string
 	}{
 		{
 			name:         "counter",
@@ -132,6 +137,63 @@ func TestPrometheusExporter(t *testing.T) {
 				histogram.Record(ctx, 23, attrs...)
 			},
 		},
+		{
+			name:          "empty resource",
+			emptyResource: true,
+			expectedFile:  "testdata/empty_resource.txt",
+			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
+				attrs := []attribute.KeyValue{
+					attribute.Key("A").String("B"),
+					attribute.Key("C").String("D"),
+					attribute.Key("E").Bool(true),
+					attribute.Key("F").Int(42),
+				}
+				counter, err := meter.SyncFloat64().Counter("foo", instrument.WithDescription("a simple counter"))
+				require.NoError(t, err)
+				counter.Add(ctx, 5, attrs...)
+				counter.Add(ctx, 10.3, attrs...)
+				counter.Add(ctx, 9, attrs...)
+			},
+		},
+		{
+			name: "custom resource",
+			customResouceAttrs: []attribute.KeyValue{
+				attribute.Key("A").String("B"),
+				attribute.Key("C").String("D"),
+			},
+			expectedFile: "testdata/custom_resource.txt",
+			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
+				attrs := []attribute.KeyValue{
+					attribute.Key("A").String("B"),
+					attribute.Key("C").String("D"),
+					attribute.Key("E").Bool(true),
+					attribute.Key("F").Int(42),
+				}
+				counter, err := meter.SyncFloat64().Counter("foo", instrument.WithDescription("a simple counter"))
+				require.NoError(t, err)
+				counter.Add(ctx, 5, attrs...)
+				counter.Add(ctx, 10.3, attrs...)
+				counter.Add(ctx, 9, attrs...)
+			},
+		},
+		{
+			name:              "without target_info",
+			withoutTargetInfo: true,
+			expectedFile:      "testdata/without_target_info.txt",
+			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
+				attrs := []attribute.KeyValue{
+					attribute.Key("A").String("B"),
+					attribute.Key("C").String("D"),
+					attribute.Key("E").Bool(true),
+					attribute.Key("F").Int(42),
+				}
+				counter, err := meter.SyncFloat64().Counter("foo", instrument.WithDescription("a simple counter"))
+				require.NoError(t, err)
+				counter.Add(ctx, 5, attrs...)
+				counter.Add(ctx, 10.3, attrs...)
+				counter.Add(ctx, 9, attrs...)
+			},
+		},
 	}
 
 	for _, tc := range testCases {
@@ -139,7 +201,12 @@ func TestPrometheusExporter(t *testing.T) {
 			ctx := context.Background()
 			registry := prometheus.NewRegistry()
 
-			exporter, err := New(WithRegisterer(registry))
+			opts := []Option{WithRegisterer(registry)}
+			if tc.withoutTargetInfo {
+				opts = append(opts, WithoutTargetInfo())
+			}
+
+			exporter, err := New(opts...)
 			require.NoError(t, err)
 
 			customBucketsView, err := view.New(
@@ -152,7 +219,28 @@ func TestPrometheusExporter(t *testing.T) {
 			defaultView, err := view.New(view.MatchInstrumentName("*"))
 			require.NoError(t, err)
 
-			provider := metric.NewMeterProvider(metric.WithReader(exporter, customBucketsView, defaultView))
+			var res *resource.Resource
+
+			if tc.emptyResource {
+				res = resource.Empty()
+			} else {
+				res, err = resource.New(ctx,
+					// always specify service.name because the default depends on the running OS
+					resource.WithAttributes(semconv.ServiceNameKey.String("prometheus_test")),
+					// Overwrite the semconv.TelemetrySDKVersionKey value so we don't need to update every version
+					resource.WithAttributes(semconv.TelemetrySDKVersionKey.String("latest")),
+					resource.WithAttributes(tc.customResouceAttrs...),
+				)
+				require.NoError(t, err)
+
+				res, err = resource.Merge(resource.Default(), res)
+				require.NoError(t, err)
+			}
+
+			provider := metric.NewMeterProvider(
+				metric.WithResource(res),
+				metric.WithReader(exporter, customBucketsView, defaultView),
+			)
 			meter := provider.Meter("testmeter")
 
 			tc.recordMetrics(ctx, meter)